separate landing components to separate files.
